+++ docs/conf.py
@@ -35,7 +35,10 @@ needs_sphinx = '1.8'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = ['sphinx.ext.intersphinx',
-              'sphinx.ext.todo']
+              'sphinx.ext.todo',
+              'nbsphinx',
+              'nbsphinx_link',
+              'sphinx_gallery.load_style']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -147,7 +150,8 @@ html_favicon = "_static/immcantation.ico"
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
-html_context = {'css_files': ['_static/overrides.css']}
+#html_context = {'css_files': ['_static/overrides.css']}
+html_css_files = ['overrides.css']
 
 # Add any extra paths that contain custom files (such as robots.txt or
 # .htaccess) here, relative to this directory. These files are copied